pyproject.toml
[tool.poetry]
name = "shaman-project"
version = "0.1.0"
description = ""
authors = ["Sophie Robert <sophie.robert@atos.net>"]
packages = [
    { include = "bbo", from = "shaman_project/" },
    { include = "bb_wrapper", from = "shaman_project/" },
    { include = "shaman_api", from = "shaman_project/" },
    { include = "shaman_worker", from = "shaman_project/" }
]

[tool.poetry.dependencies]
python = "^3.6.1"
fastapi = { optional = true, version = "^0.61.1" }
uvicorn = { optional = true, version = "^0.11.8" }
arq = { optional = true, version = "^0.19" }
motor = { optional = true, version = "^2.2.0" }
pydantic = { optional = true, version = "^1.6.1" }
requests = { optional = true, version = "^2.24.0" }
typer = { optional = true, version = "^0.3.2" }
PyYAML = { optional = true, version = "^5.3.1" }
prometheus-client = { optional = true, version = "^0.8.0" }
starlette-exporter = { optional = true, version = "^0.4.1" }
httpx = { optional = true, version = "^0.14.3" }
numpy = { optional = true, version = "^1.19.2" }
scikit-learn = { optional = true, version = "^0.23.2" }
cma = { optional = true, version = "^3.0.3" }
scipy = { optional = true, version = "^1.5.2" }
pandas = { optional = true, version = "^1.1.2" }

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
invoke = "^1.4.1"
devtools = "^0.6"

[tool.poetry.extras]
bb-wrapper = ["pydantic", "requests", "typer", "PyYAML"]
bbo = ["numpy", "scikit-learn", "cma", "scipy", "pandas"]
shaman-api = ["fastapi", "uvicorn", "arq", "motor", "starlette-exporter", "typer"]
shaman-worker = ["arq", "prometheus-client", "httpx", "pydantic"]

[tool.poetry.scripts]
shaman-api = "shaman_api.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
